Allow loading of credentials from the default chain.
In the current implementation, credentials must be provided in the
pom.xml in order for the plugin to work correctly.  This requires that
they either be added to the settings.xml, or be checked into source
control for all others to see.

By allowing the default credentials chain to be used, any sytem that's
already configured to use AWS services will be able to update stacks as
well.  When a user updates their credentials through the command line
tool, they won't need to remember to update their settings.xml file.